backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
supabase>=2.3.0
clerk-backend-api>=0.5.0
resend>=0.7.0
python-dotenv>=1.0.0
pydantic>=2.5.0
httpx>=0.26.0
python-multipart>=0.0.6
email-validator>=2.1.0
